+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- Rename continue_conversation_with_tool_results to continue_chat_with_tool_results (#123)
 - Removed `tool_call: ToolCall` attr from `ToolCallResult` (#119)
 
 ## [0.0.9] - 2025-08-09